Refactor models to own modules
